+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Query size limiter.
 - Fix /threads/messages endpoint.
 - [frontend] - remove first message delay.
+- [frontend] - improve loading UI on first message.
 - [frontend] - MarkDown rendering.
 - [frontend] - dark mode flickering.
 - [frontend] - Concatenate base url and endpoint correctly.